Refactor Harvard date parsing
